requirements.txt
fastapi[all]==0.111.0
rdflib
pyld
pydantic-settings
pyshacl
ogc-na @ git+https://github.com/opengeospatial/ogc-na-tools@main#egg=ogc-na
